Moved async fs functions that return promises to a utility file to be imported into index.js. #3
