+++ server/index.js
@@ -8,7 +8,7 @@ const io = require('socket.io').listen(server)
 const bodyParser = require('body-parser')
 const jsonParser = bodyParser.json()
 
-const fs = require('fs')
+const { readFileAsync, writeFileAsync, readdirAsync, unlinkAsync } = require('./utils/fsAsyncFunctions')
 
 let currentCanvas = null
 
@@ -62,40 +62,4 @@ function newConnection(socket) {
     socket.broadcast.emit('mouse', data)
     console.log(data)
   }
-}
-
-function readFileAsync(fileName) {
-  return new Promise((resolve, reject) => {
-    fs.readFile(fileName, 'utf8', (err, data) => {
-      if (err) return reject(err)
-      return resolve(data)
-    })
-  })
-}
-
-function writeFileAsync(fileName, data) {
-  return new Promise((resolve, reject) => {
-    fs.writeFile(fileName, data, 'utf8', (err) => {
-      if (err) return reject(err)
-      else return resolve()
-    })
-  })
-}
-
-function readdirAsync(path) {
-  return new Promise((resolve, reject) => {
-    fs.readdir(path, (err, files) => {
-      if (err) return reject(err)
-      return resolve(files)
-    })
-  })
-}
-
-function unlinkAsync(path) {
-  return new Promise((resolve, reject) => {
-    fs.unlink(path, (err) => {
-      if (err) return reject(err)
-      return resolve()
-    })
-  })
 }
